# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Calibrate API
  version: 1.0.0
  description: Stable wire contract shared by the self-hosted web, Android, and future Wear OS clients.
servers:
  - url: /
paths:
...
  /auth/mobile/login:
    post:
      operationId: loginMobile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileAuthRequest'
      responses:
        '200':
          description: Authenticated native session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mobile/register:
    post:
      operationId: registerMobile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileAuthRequest'
      responses:
        '200':
          description: Registered Android phone session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAuthResponse'
        '400':
          description: Invalid registration or Android phone device metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /auth/mobile/refresh:
    post:
      operationId: refreshMobile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Rotated native session credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileRefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mobile/sessions:
    get:
      operationId: listMobileSessions
      responses:
        '200':
          description: Active native sessions owned by the account.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/MobileSessionSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/mobile/wear/pairing-credential:
    post:
      operationId: issueWearPairingCredential
      description: Issue a short-lived one-time credential from the current Android phone session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WearPairingCredentialRequest'
      responses:
        '201':
          description: Credential and exact self-hosted server origin to transfer to the paired watch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WearPairingCredentialResponse'
        '400':
          description: Invalid server origin, watch identity, protocol, or public key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not an active Android phone session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /auth/mobile/wear/pair:
    post:
      operationId: exchangeWearPairingCredential
      security: []
      description: Consume a server-bound one-time credential and create a revocable Wear OS session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WearPairingExchangeRequest'
      responses:
        '200':
          description: Authenticated Wear OS session credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WearMobileAuthResponse'
        '400':
          description: Invalid exchange payload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401':
          description: Credential or challenge signature is invalid, or the issuing phone session is inactive.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409':
          description: Credential was already used, its binding differs, or a same-exchange UUID retry within the credential TTL revoked the orphaned session and requires fresh pairing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '410':
          description: Credential expired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
components:
  securitySchemes:
    mobileBearer:
      type: http
      scheme: bearer
...
  responses:
    Unauthorized:
      description: Authentication is missing, expired, or rejected.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: [string, 'null']
        field_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        retryable:
          type: boolean
...
    MobileAuthRequest:
      type: object
      required: [email, password, device_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
        device_id:
          type: string
          maxLength: 128
        device_platform:
          type: string
          enum: [android_phone]
        device_name:
          type: string
          maxLength: 120
    MobileAuthResponse:
      type: object
      required: [user, access_token, refresh_token, access_expires_at, refresh_expires_at]
      properties:
        user:
          $ref: '#/components/schemas/UserClientPayload'
        access_token:
          type: string
        refresh_token:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
    WearAuthPrincipal:
      type: object
      additionalProperties: false
      required: [id, timezone, language, weight_unit, height_unit]
      properties:
        id:
          type: integer
        timezone:
          type: string
        language:
          type: string
        weight_unit:
          type: string
          enum: [KG, LB]
        height_unit:
          type: string
          enum: [CM, FT_IN]
    WearMobileAuthResponse:
      type: object
      required: [user, access_token, refresh_token, access_expires_at, refresh_expires_at]
      properties:
        user:
          $ref: '#/components/schemas/WearAuthPrincipal'
        access_token:
          type: string
        refresh_token:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
    MobileRefreshResponse:
      oneOf:
        - $ref: '#/components/schemas/MobileAuthResponse'
        - $ref: '#/components/schemas/WearMobileAuthResponse'
    MobileSessionSummary:
      type: object
      required:
        - id
        - device_id
        - device_platform
        - device_name
        - created_at
        - last_used_at
        - refresh_expires_at
        - current
      properties:
        id:
          type: integer
        device_id:
          type: string
        device_platform:
          $ref: '#/components/schemas/MobileDevicePlatform'
        device_name:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: [string, 'null']
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
        current:
          type: boolean
    WearPairingCredentialRequest:
      type: object
      additionalProperties: false
      required: [server_origin, watch_device_id, protocol_version, watch_public_key_spki]
      properties:
        server_origin:
          type: string
          format: uri
          description: Exact Calibrate origin already selected and validated by the phone.
        watch_device_id:
          type: string
          minLength: 1
          maxLength: 128
        watch_device_name:
          type: string
          maxLength: 120
        protocol_version:
          type: integer
          enum: [1]
        watch_public_key_spki:
          type: string
          description: Base64 DER SubjectPublicKeyInfo for a P-256 key held by Android Keystore.
    WearPairingCredentialResponse:
      type: object
      additionalProperties: false
      required: [pairing_token, server_origin, watch_device_id, protocol_version, challenge, expires_at]
      properties:
        pairing_token:
          type: string
        server_origin:
          type: string
          format: uri
        watch_device_id:
          type: string
        protocol_version:
          type: integer
          enum: [1]
        challenge:
          type: string
          description: Base64url server nonce the watch signs with SHA256withECDSA.
        expires_at:
          type: string
          format: date-time
    WearPairingExchangeRequest:
      type: object
      additionalProperties: false
      required: [pairing_token, server_origin, watch_device_id, protocol_version, exchange_id, challenge_signature]
      properties:
        pairing_token:
          type: string
          minLength: 1
          maxLength: 256
        server_origin:
          type: string
          format: uri
        watch_device_id:
          type: string
          minLength: 1
          maxLength: 128
        protocol_version:
          type: integer
          enum: [1]
        exchange_id:
          type: string
          format: uuid
          description: Client-generated UUID retained unchanged when retrying an exchange after response loss.
        challenge_signature:
          type: string
          description: Base64url DER SHA256withECDSA signature over the protocol v1 payload, including exchange_id.
...
    MobileDevicePlatform:
      type: string
      enum: [android_phone, wear_os]
...
    UserClientPayload:
      type: object
      required:
        - id
        - email
        - created_at
        - weight_unit
        - height_unit
        - timezone
        - language
        - reminder_log_weight_enabled
        - reminder_log_food_enabled
        - haptics_enabled
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        weight_unit:
          type: string
          enum: [KG, LB]
        height_unit:
          type: string
          enum: [CM, FT_IN]
        timezone:
          type: string
        language:
          type: string
        reminder_log_weight_enabled:
          type: boolean
        reminder_log_food_enabled:
          type: boolean
        haptics_enabled:
          type: boolean
        date_of_birth:
          type: [string, 'null']
          format: date
        sex:
          type: [string, 'null']
          enum: [MALE, FEMALE, null]
        height_mm:
          type: [integer, 'null']
        activity_level:
          type: [string, 'null']
          enum: [SEDENTARY, LIGHT, MODERATE, ACTIVE, VERY_ACTIVE, null]
        profile_image_url:
          type: [string, 'null']
security:
  - mobileBearer: []